arch: migrate science coordination runtimes to surface shims
